backend/pyproject.toml
[tool.poetry]
name = "vcgpt-backend"
version = "0.1.0"
description = ""
authors = ["GHuiKai <aodesai19899278@13.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
langchain = "^0.0.344"
llama-index = "^0.9.10"
openai = "^1.3.7"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
python-dotenv = "^1.0.0"
duckduckgo-search = "^3.9.9"
numexpr = "^2.8.7"
langchainhub = "^0.1.14"
langserve = {extras = ["all"], version = "^0.0.32"}
pydantic = "1.10.13"
streamlit = "^1.29.0"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
